requirements.txt
psycopg2_binary
Django==3.1.4
pymysql
django-environ==0.4.5
prometheus_client
retry.it
cachetools
djangorestframework-filters==1.0.0.dev2
djangorestframework-xml
djangorestframework==3.12.4
djangorestframework-gis==0.17
django-utils-six==2.0
django-filter==21.1
tqdm
multiprocessing-logging
requests
pyyaml